[BUG] Skip Empty Partitions in Bulk Sample Writing (#3600)
Currently, when the bulk sampler writes its samples, it does not check whether the partition is empty.  This causes all of the empty partitions to collide and attempt to write to the same (empty) file at once.  This writing should not occur at all.

This PR adds a check that skips writing if the partition is empty.

This should resolve the nightly test failures.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Vibhu Jawa (https://github.com/VibhuJawa)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/3600