[FIX] Add mkl version, limit tensordict to 0.6.2 (#161)
Adds a constraint on the `mkl` and `tensordict` versions to prevent
installing incompatible versions of PyTorch, as well as incompatible
versions of `tensordict`. We were previously installing the cpu-only
version of PyTorch in nighly tests and tensordict 0.7.2, which appears
to be incompatible with *some* versions of PyTorch 2.6 (version 0.6.2
works without issue).