[BUG] Specify Input Type and Assign Output to Correct Input Type (#236)
**Fixes Two Major Bugs**

Fixes a major bug that resulted in output data being assigned to the wrong input type under certain circumstances.  The input type is now passed as metadata to the sampler so it can be retrieved after the sampling process.  This ensures `num_sampled_nodes` and `num_sampled_edges` are filled appropriately when using the `LinkNeighborLoader`.  It also ensures that `input_id` is assigned to the correct vertex or edge type.

Fixes a second major bug where the output edge inverse was not properly de-offsetted, causing invalid output.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Tingyu Wang (https://github.com/tingyu66)

URL: https://github.com/rapidsai/cugraph-gnn/pull/236