[BUG] Fix Calculation of # Sampled Nodes/Edges with Zero Input Size (#283)
Fixes a bug found when running `movielens_mnmg.py` on input with a large number of negative edges.  Some batches end up with no positive edges of the input type, which causes PyTorch to throw an exception when trying to calculate the number of nodes, since the returned tensor is empty.

The fix is to check if the returned tensor is empty, and just set the number of output nodes to zero if it is.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Tingyu Wang (https://github.com/tingyu66)

URL: https://github.com/rapidsai/cugraph-gnn/pull/283