Pin Cython pre-3.2.0 and PyTest pre-9 (#345)
Pin to workaround for issues:
* https://github.com/rapidsai/build-planning/issues/229
* https://github.com/rapidsai/build-planning/issues/230

## Description

Recently Cython 3.2.0 was released and we have seen a few subtle issues
building with it. While we work out these issues, this pins to Cython
3.1, which know to be working for us.

Similarly PyTest 9 was recently released, but we have ran into some
issues with it as well. So pin to PyTest 8 while we work through PyTest
9 issues.

---------

Co-authored-by: Alex Barghi <105237337+alexbarghi-nv@users.noreply.github.com>
Co-authored-by: Gil Forsyth <gforsyth@users.noreply.github.com>